analytics page

+++ app/src/main/java/com/example/tracktor/data/model/InventoryItem.kt
@@ -3,7 +3,7 @@ package com.example.tracktor.data.model
 data class InventoryItem (
     val userStats: MutableMap<String,UserInventoryStat> = mutableMapOf<String,UserInventoryStat>(),
     val name: String = "",
-    val itemTotal : Int = 0,
+    val itemTotal : Long = 0,
     val itemPrice: Double = 0.0,
     val imageRef: String? = null,
 )
